Second optimal packer (partial)
Alteration of the optimal rectangle packer.

I did not like the way the placement matrix was indexed, so I changed
that. To fix the bug from the previous commit, I put no restriction on
the size of the placement matrix in the greedy solver. Then every time a
new bounding box is checked, a new placement matrix is generated.

Rewritten the code as to make its structure clearer based on the paper
KORF 2010. I could not see the connection with some code segments of the
previous optimal packer and the paper.

Finally, the previous packer contained some code duplication, for
example, the methods  anytimeSolution  and  anytimeSolutionFixedHeight
were very similair, so these are combined in one method: anytimeSolution