Pasted in pruning on wasted space
Pasted in pruning based on wasted space. There are some remarks I'd like
to make however.. (i) We only computed the wasted space on empty rows,
adding pruninng on empty columns could still be added. (ii) We have to
recalculate the histogram each time, whilst this could be done
iteratively. Korf 2010 also mentions that ,,is the main reason for the
large performance
improvement between the results reported here and those in Korf
(2004).''