Implemented different searching order
If the height is not fixed, two greedy solutions are produced. Whichever
one has the least amount of area is taken as a starting point.

If the wide greedy solution is used, the bounding box is made less wide
but taller. If the tall greedy solution is used, the bounding box is
made less tall but wider.