Some name changes for clarity
There are two pruning techniques with dominance conditions used in the
paper, both use areas of empty space, so I thought this name not
fitting. Instead it is described with the terms 'PerfectRectangles' as
opposed to the other technique with narrow strips of empty space.

The names for the functions canPruneDominanceEmptySpaceLeft etc are left
as they are as it is not entirely clear to me if these are supposed to
also contain the different pruning technique in the future, in which
case it might be best to ditch the 'empty space' part.