Fixed small bug dominance pruning + Adde simplification with gcd
There was a small fault in the implementation of the dominance pruning
with perfect rectangles: if the rectangle was perfect but contained no
empty space, it was pruned whilst it should not hve been pruned.

Added simplification with gcd per dimension. This also implied changeing
the fixedHeight and the totalRectArea during the computation.

I suggest we also use this method for the other algorithms, but we'll
see.